+++ src/main/prim/_breed.scala
@@ -1,9 +1,9 @@
 // (C) Uri Wilensky. https://github.com/NetLogo/NetLogo
 
-package org.nlogo.prim.etc
+package org.nlogo.prim
 
 import org.nlogo.agent.AgentSet
-import org.nlogo.core.{ Syntax }
+import org.nlogo.core.Syntax
 import org.nlogo.nvm.{ Reporter, Context }
 
 class _breed(breedName: String) extends Reporter {
