+++ src/main/prim/_inradius.scala
@@ -1,6 +1,6 @@
 // (C) Uri Wilensky. https://github.com/NetLogo/NetLogo
 
-package org.nlogo.prim.etc
+package org.nlogo.prim
 
 import
   java.util.{ List => JList }
